+++ packages/orq-rc/src/orq_ai_sdk/models/createeval_function_params_evals_response_200_applicationjson_responsebody_5_type.py
@@ -1650,9 +1650,9 @@ class Typescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-05-20T05:21:04.498Z"
+    created: Optional[str] = "2026-05-20T07:16:36.761Z"
 
-    updated: Optional[str] = "2026-05-20T05:21:04.498Z"
+    updated: Optional[str] = "2026-05-20T07:16:36.761Z"
 
     guardrail_config: OptionalNullable[
         CreateEvalResponseBodyEvalsResponse200ApplicationJSONGuardrailConfig
@@ -1902,9 +1902,9 @@ class Ragas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-05-20T05:21:04.498Z"
+    created: Optional[str] = "2026-05-20T07:16:36.761Z"
 
-    updated: Optional[str] = "2026-05-20T05:21:04.498Z"
+    updated: Optional[str] = "2026-05-20T07:16:36.761Z"
 
     guardrail_config: OptionalNullable[
         CreateEvalResponseBodyEvalsResponse200GuardrailConfig