+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS1X9HFSRR4ZZJ7XWAPSC7NG"
+        "01KS23X26JVB3ABV0ZB1KQKZX5"
     )
     r"""The unique identifier of the data source"""