+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS1X9HFWBAVVG00BC79Y3GGJ"
+        "01KS23X26NNM66M12YJ7HTGPAX"
     )
     r"""The unique identifier of the data source"""