+++ packages/orq-rc/RELEASES.md
@@ -8049,3 +8049,13 @@ Based on:
 - [python v4.10.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.19 - packages/orq-rc
+
+## 2026-05-20 07:19:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.3 (2.884.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.20 - packages/orq-rc